+++ client/app/services/backendService.js
@@ -10,18 +10,18 @@
     function backendService($http) {
 
         return {
-            getModules: getModules,
-            getTimeline: getTimeline,
-            getReadme: getReadme,
-            getUsersProfile: getUsersProfile,
-            updateUserRole: updateUserRole
+            getModules,
+            getTimeline,
+            getReadme,
+            getMyProfile,
+            getUsersProfile,
+            updateUserRole
         }
 
         function getModules() {
             return $http.get('/api/modules', getHttpConfig())
                 .then(res => res.data);
         }
-
 
         function getTimeline() {
             return $http.get('/api/groups')
@@ -30,6 +30,11 @@
 
         function getReadme(gitRepo) {
             return $http.get('/api/github/readme/hackyourfuture/' + gitRepo)
+                .then(res => res.data);
+        }
+
+        function getMyProfile() {
+            return $http.get('/api/user', getHttpConfig())
                 .then(res => res.data);
         }